Create GSwP- 19 of 28
if ("Jon".length * 2 / (2+1) === 2)
{
    console.log("The answer makes sense!");
} 
else
{
    console.log("Error Error Error");
}



/*


I added an if in the first line to make the if/ else statement.

The === operator means "is exactly equal to," matching by both value and data type.

The == operator means "is equal to," matching by value only.

So the length of the text string is 3 * 2 = 6  divided by  3 (2+1)

so 6 / 3 === 2

So I inputted the value 2

I then added


{
    console.log("Error Error Error");
}


to complete the else statement.



Original code:


 ("Jon".length * 2 / (2+1) === )
{
    console.log("The answer makes sense!");
} 
else 


Instructions
Complete the missing bits of code to construct the if / else statement. Make the condition evaluate to true.
Finish the else statement by printing out the string "Error Error Error" to the console.



"Math

We saw basic math before. The basic math symbols we learned in school work here. Even the order in which the computer understands the math is the same as in school!

Code:

1. ( ): control order of operations
2. * and /: multiplication and division
3. - and +: subtraction and addition

Examples:

1. 100/10 evaluates to 10
2. "Jane".length + 5 evaluates to 9
3. 5*(3+1) evaluates to 20"


*/