Create GSwP- 20 of 28
console.log(14 % 3)
console.log(99 % 8)
console.log(11 % 3)


/*

The Output of this is:

2
3
2


14 / 3 = 4  Remainder 2
99 / 8 = 96 Remainder 3
11 / 3 = 9  Remainder 2






Original code

// Below is an example of printing the remainder of 18/4 using modulo:
// console.log(18 % 4); 


Math and the modulo

Let's meet an interesting symbol called modulo. When % is placed between two numbers, the computer will divide the first number by the second, and then return the remainder of that division.

So if we do 23 % 10, we divide 23 by 10 which equals 2 with 3 left over. So 23 % 10 evaluates to 3.

More examples:

17 % 5 evaluates to 2

13 % 7 evaluates to 6

Instructions
Use console.log and modulo three times to print the remainder of the following equations:

a. 14 / 3
b. 99 / 8
c. 11 / 3


*/